floating point calculations

+++ nav2_mppi_controller/src/critics/path_align_critic.cpp
@@ -89,7 +89,7 @@ void PathAlignCritic::score(CriticData & data)
   }
 
   float dist_sq = 0, dx = 0, dy = 0, dyaw = 0, summed_dist = 0;
-  double min_dist_sq = std::numeric_limits<float>::max();
+  float min_dist_sq = std::numeric_limits<float>::max();
   size_t min_s = 0;
 
   for (size_t t = 0; t < batch_size; ++t) {
@@ -118,7 +118,7 @@ void PathAlignCritic::score(CriticData & data)
       // The nearest path point to align to needs to be not in collision, else
       // let the obstacle critic take over in this region due to dynamic obstacles
       if (min_s != 0 && (*data.path_pts_valid)[min_s]) {
-        summed_dist += std::sqrt(min_dist_sq);
+        summed_dist += sqrtf(min_dist_sq);
       }
     }
